mini-update idt.c

+++ kernel/cpu/idt.c
@@ -11,7 +11,7 @@ void set_idt_gate(int n, uint32_t handler)
 
 void set_idt()
 {
-    idt_reg.base = (uint32_t)&idt;
+    idt_reg.base = (uint64_t)&idt;
     idt_reg.limit = IDT_ENTRIES * sizeof(idt_gate_t) - 1;
     __asm__ __volatile__ ("lidtl (%0)" : : "r" (&idt_reg));
 }
